+++ contracts/protocol/interface/IIFlareSystemsCalculator.sol
@@ -19,7 +19,7 @@ interface IIFlareSystemsCalculator is IFlareSystemsCalculator {
      */
     function calculateRegistrationWeight(
         address _voter,
-        uint24 _rewardEpochId,
+        uint32 _rewardEpochId,
         uint256 _votePowerBlockNumber
     )
         external
